modal/close/revise
* 카드를 삭제할때 confirm 받기
* confirm 창을 띄워 확인 또는 취소 버튼을 누르게 하고 취소를 누르면 종료, 확인을 누르면 해당 카드를 삭제하는 로직 추가
* confirm에 들어가는 문장은 confirmSentence라는 변수에 넣어둠

+++ frontend/src/modules/deleteTodo.js
@@ -1,10 +1,14 @@
 import { deleteFetchManager } from './fetchManger';
 import { updateCount } from './utils/updateCount.js';
+const confirmSentence = '선택하신 카드를 삭제하겠습니까?';
 export default async function deleteTodo(e) {
   if (e.target.dataset.method !== 'delete') return;
+  if (!checkConfirm()) return;
+
   const deleteItem = e.target.closest('li');
   const id = deleteItem.id;
   const ul = e.target.closest('ul');
+
   try {
     const result = await deleteFetchManager(`/api/todo/${id}`);
     if (result.status !== 200) throw new Error();
@@ -13,4 +17,9 @@ export default async function deleteTodo(e) {
   } catch (e) {
     alert('다시 시도해주세요');
   }
+}
+
+function checkConfirm() {
+  const result = confirm(confirmSentence);
+  return result;
 }
